requirements.txt
absl-py==2.1.0
accelerate==0.31.0
einops~=0.8.1
evaluate==0.4.3
faiss_cpu; sys_platform == 'darwin'
faiss_gpu==1.7.2; sys_platform != 'darwin'
hydra-core==1.3.2
huggingface-hub<0.26.0 # Required because another dependency requires the "cached_download" function
joblib~=1.4.2
langchain==0.3.23
langchain-community==0.3.15
langchain-huggingface~=0.1.2
langchain-ollama~=0.3.2
langchain-openai~=0.3.13
langgraph==0.2.62
nltk==3.8.1
numpy<2.0.0
numpyencoder==0.3.0
openai>=1.61.0
pandas~=2.2.3
Requests==2.32.3
rouge_score==0.1.2
scispacy==0.5.5
semanticscholar==0.9.0
tiktoken==0.7.0
tinydb==4.8.0
torch>=2.1.0
toml==0.10.2
transformers>=4.47.1
uplink~=0.9.7
urllib3<2.0.0
vllm==0.7.2 # NOTE: This requires CUDA to run, and is required for running the LLM evaluation.
prometheus-eval~=0.1.20 # NOTE: Remove this if CUDA is not available